+++ src/io-schema.ts
@@ -1,22 +1,31 @@
 import { z } from "zod";
+import { FlatObject } from "./common-helpers";
 import { copyMeta } from "./metadata";
 import { AbstractMiddleware } from "./middleware";
 import { RawSchema } from "./raw-schema";
 
-type Refined<T extends z.ZodTypeAny> =
-  T extends z.ZodType<infer O> ? z.ZodEffects<T | Refined<T>, O, O> : never;
+type BaseObject<U extends z.UnknownKeysParam> = z.ZodObject<z.ZodRawShape, U>;
+
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type -- workaround for TS2456, circular reference
+interface ObjectBasedEffect<T extends z.ZodTypeAny>
+  extends z.ZodEffects<T, FlatObject> {}
+
+type EffectsChain<U extends z.UnknownKeysParam> = ObjectBasedEffect<
+  BaseObject<U> | EffectsChain<U>
+>;
 
 /**
  * @desc The type allowed on the top level of Middlewares and Endpoints
  * @param U — only "strip" is allowed for Middlewares due to intersection issue (Zod) #600
  * */
 export type IOSchema<U extends z.UnknownKeysParam = z.UnknownKeysParam> =
-  | z.ZodObject<z.ZodRawShape, U>
-  | z.ZodUnion<[IOSchema<U>, ...IOSchema<U>[]]>
-  | z.ZodIntersection<IOSchema<U>, IOSchema<U>>
-  | z.ZodDiscriminatedUnion<string, z.ZodObject<z.ZodRawShape, U>[]>
-  | Refined<z.ZodObject<z.ZodRawShape, U>>
-  | RawSchema;
+  | BaseObject<U> // z.object()
+  | EffectsChain<U> // z.object().refine(), z.object().transform(), z.object().preprocess()
+  | RawSchema // ez.raw()
+  | z.ZodUnion<[IOSchema<U>, ...IOSchema<U>[]]> // z.object().or()
+  | z.ZodIntersection<IOSchema<U>, IOSchema<U>> // z.object().and()
+  | z.ZodDiscriminatedUnion<string, BaseObject<U>[]> // z.discriminatedUnion()
+  | z.ZodPipeline<ObjectBasedEffect<BaseObject<U>>, BaseObject<U>>; // z.object().remap()
 
 /**
  * @description intersects input schemas of middlewares and the endpoint